requirements.txt
# TTS Subtitling Studio - Dependencies

# UI Framework
flet>=0.25.0

# ONNX Runtime for TTS inference
onnxruntime>=1.18.0

# Audio processing
soundfile>=0.12.0
pydub>=0.25.0

# Gemini API
google-genai>=0.8.0

# Utilities
numpy>=1.24.0